refactor: extract lib.rs tests + presenter settings helper to stay under file-size cap
Phase 1+2 commits pushed lib.rs to 1012 lines (cap is 1000). Two changes:

1. Move the `#[cfg(test)] mod tests` block (lines 650-1007) into
   crates/sp-server/src/lib_tests.rs. lib.rs now declares
   `#[cfg(test)] mod lib_tests;` and includes nothing else about
   tests. Test coverage unchanged; cargo test still picks them up.

2. Move the presenter_url / presenter_enabled settings read + client
   build out of lib.rs into a new `presenter::build_from_settings`
   helper. lib.rs now reads:
     let presenter_client = presenter::build_from_settings(\&pool).await?;
   instead of a 20-line block.

Net: lib.rs drops from 1032 to 651. Well under the 1000-line cap
and leaves room for future additions.
